Revert "Verval beta uitbreiding"
