--- Assets/Scripts/Manager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: ed0b34eb6ba98314e8cb84aff5ef3893
