--- Assets/StreamingAssets.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 00b0fa12453b08d478f0b3c73b76f15e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
