--- Assets/StreamingAssets/VervalBetaData.csv
@@ -1,4 +1,0 @@
-t(s),hits
-10,237
-20,210
-30,208
