--- Assets/StreamingAssets/VervalBetaData.csv.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: da5c5895ecb09524d8057dab490ffd37
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
